requirements.txt
discord.py
requests
gmail
